requirements.txt
tabulate
python-dotenv
colorama